+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 v2.2.0 (working):
+* [ISSUE#18] Default user creation error on Alpine
 * [BUG] start action on alpine not working
 * [EvVOL] Manage wsl output list as object (no more strings)
 * [ISSUE#19] Display error wslctl ls when no distributions